Respect windowFeatures with falsy windowName
Currently, if you do `blankshield.open('url', '', 'width=100')` the `'width=100'` (and any other windowFeatures) is effectively ignored because they actually get passed to `window.open` as the second argument (windowName). This change always fills in a windowName argument, so that windowFeatures will always be correctly passed as the third argument.